More fixing up naming
